refactoring the AActionHandler functions to accept the calling action queue as a parameter; extending sorting to accept exceptions like in 'sort +zl $B $c'; making sure that all macros in xmacro.cpp get only the atoms of the model in focus - excluding the overlayed models;

+++ glib/gltextbox.cpp
@@ -172,7 +172,7 @@ void TGlTextBox::Fit()  {
         tr.width *= scale;
       }
       Height += (uint16_t)olx_round(olx_max(tr.height, glf.GetMaxHeight()));
-      if (tr.width > Width) Width = olx_round(tr.width);
+      if (tr.width > Width) Width = (uint16_t)olx_round(tr.width);
     }
     Height += (uint16_t)olx_round(LineSpacer*(FBuffer.Count()-1));
   }
